Create Function.kt
Return a value from a function
In more complex apps, functions do more than print text.

Kotlin functions can also generate data called a return value which is stored in a variable that you can use elsewhere in your code.

When defining a function, you can specify the data type of the value you want it to return. The return type is specified by placing a colon (:) after the parentheses, a single blank space, and then the name of the type (Int, String, etc). A single space is then placed between the return type and the opening curly brace. Within the function body, after all the statements, you use a return statement to specify the value you want the function to return. A return statement consists of the return keyword followed by the value, such as a variable, you want the function to return as an output.